requirements.txt
Eel
pythonping
speedtest-cli
requests